web-animations/timing-model/animations/canceling-an-animation.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Canceling an animation: cancel event</title>
<link rel="help" href="https://w3c.github.io/web-animations/#canceling-an-animation-section">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>

async_test(function(t) {
  var anim1 = createDiv(t).animate(null, 100 * MS_PER_SEC);
  var anim2 = createDiv(t).animate(null, 100 * MS_PER_SEC);

  var fired = false;

  anim1.oncancel = function(event) {
    t.step(function() {
      assert_equals(event.target, anim1, 'Target of cancel event should be anim1.');
      assert_equals(event.currentTime, null, 'currentTime of cancel event should be null.');
      assert_equals(event.timelineTime, document.timeline.currentTime, 'Event timelineTime should be same as document.timeline.currentTime.');
    });
    fired = true;
  };

  anim2.onfinish = function() {
    t.step(function() {
      assert_true(fired, 'anim1.oncancel should be called.');
    });
    t.done();
  };

  anim1.cancel();
  anim2.finish();
}, 'Cancelling animation should fire a cancel event.');

</script>
</body>
